requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-core
langgraph
requests
python-dotenv
